Update README.md
While reviewing for https://github.com/openjournals/jose-reviews/issues/259 I found a few changes to suggest:

I added a few hyperlinks to local files - however it is done, I think that when files are referenced it should be clear to the reader where to find them - even if they are in the root directory, but especially if not.  

Should Rstudio be listed as a requirement? The first to methods under getting the code assume it is installed. 

This also fixes a few typos